This is part of the process of adding a mapper to marlpde.py. This mapper of Python parameter names to FORTRAN parameter names will help to derive a new dataclass from Scenario. This derived dataclass will have all the required parameters to be fed into LMAHeureuxPorosityDiff. We can then do eq = LMAHeureuxPorosityDiff(**derived_dataclass()). Of course, LMAHeureuxPorosityDiff needs a few extra fields, i.e. AragoniteSurface, CalciteSurface, CaSurface, CO3Surface, PorSurface, not_too_shallow and not_too_deep, but we can supply those as positional arguments, i.e. before the *derived_dataclass() keyword arguments.
