+++ Evolve_scenario.py
@@ -7,62 +7,7 @@ from pde.grids.operators.cartesian import _make_derivative
 import numpy as np
 import matplotlib.pyplot as plt
 from LHeureux_model import LMAHeureuxPorosityDiff
-from marlpde.marlpde import Scenario, Solver
-
-def read_Scenario_parameters():
-    '''
-    '''
-    Scenario_parameters = Scenario()
-    
-    KA = Scenario_parameters.Ka.magnitude
-    KC = Scenario_parameters.Kc.magnitude
-    CA0 = Scenario_parameters.cara0.magnitude
-    CAIni = Scenario_parameters.cara00.magnitude
-    CC0 = Scenario_parameters.ccal0.magnitude
-    CCIni = Scenario_parameters.ccal00.magnitude
-    cCa0 = Scenario_parameters.ca0.magnitude/np.sqrt(KC)
-    cCaIni = Scenario_parameters.ca00.magnitude/np.sqrt(KC)
-    cCO30 = Scenario_parameters.co30.magnitude/np.sqrt(KC)
-    cCO3Ini = Scenario_parameters.co300.magnitude/np.sqrt(KC)
-    Phi0 = Scenario_parameters.phi0.magnitude
-    PhiIni = Scenario_parameters.phi00.magnitude
-    
-    ShallowLimit = Scenario_parameters.xdis.magnitude
-    
-    DeepLimit = ShallowLimit + Scenario_parameters.Th.magnitude
-    
-    sedimentationrate = Scenario_parameters.S.magnitude
-    m1 = Scenario_parameters.m.magnitude
-    m2 = m1
-    n1 = Scenario_parameters.nn.magnitude
-    n2 = n1
-    rhoa = Scenario_parameters.rhoa.magnitude
-    rhoc = Scenario_parameters.rhoc.magnitude
-    rhot = Scenario_parameters.rhot.magnitude
-    rhos0 = rhoa * CA0 + rhoc * CC0 + rhot * (1 - (CA0 + CC0))
-    
-    rhos = rhos0
-    
-    rhow = Scenario_parameters.rhow.magnitude
-    beta = Scenario_parameters.beta.magnitude
-    D0Ca = Scenario_parameters.D0ca.magnitude
-    k1 = Scenario_parameters.k1.magnitude
-    k2 = Scenario_parameters.k2.magnitude
-    k3 = Scenario_parameters.k3.magnitude
-    k4 = Scenario_parameters.k4.magnitude
-    muA = Scenario_parameters.mua.magnitude
-    DCa = Scenario_parameters.D0ca.magnitude
-    DCO3 = Scenario_parameters.D0co3.magnitude
-    b = Scenario_parameters.b.magnitude/1e4
-    PhiNR = Scenario_parameters.phi00.magnitude
-    PhiInfty = Scenario_parameters.phiinf.magnitude
-    
-    Xstar = D0Ca / sedimentationrate
-    Tstar = Xstar / sedimentationrate 
-    
-    max_depth = Scenario_parameters.length.magnitude
-
-    return all_Scenario_parameters
+from marlpde import marlpde
 
 def read_solver_parameters():
     '''
